Added ServerSimplifiedUI.cs and ServerSimplifiedPlayer.cs and a new GameObject SimplifiedServerUI
ServerSimplifiedUI.cs is attached to SimplifiedServerUI GameObject. The gameObject should be placed in scene. This adds a GUI overlay with buttons allowing for quick and easy server creation and joining

ServerSimplifiedPlayer.cs is added to prefab Player. This works with the abovementioned. Client now moves by pressing GUI button "Move".